Make both `TypeAnnotation` and `VarAnnotation` classes available
